Introduce Course, Section, and Lesson with user progress components. Enhance error handing. Visual improvements.

+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/courseEditing/presentation/screen/CourseEditingScreen.kt
@@ -30,7 +30,7 @@ fun CourseEditingScreen(
     courseLocale: String,
     onLocaleChange: (String) -> Unit,
     sections: List<SectionDraft>,
-    onSectionClick: (Long) -> Unit,
+    onSectionClick: (Int) -> Unit,
     onSaveButtonClick: () -> Unit
 ) {
     ScreenContainer(
