+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/lesson/domain/model/LessonDraft.kt
@@ -1,25 +1,25 @@
 package cz.cvut.docta.lesson.domain.model
 
 sealed class LessonDraft(
-    open val id: Long,
+    open val id: Int,
     open val name: String
 ) {
 
     data class Default(
-        override val id: Long,
+        override val id: Int,
         override val name: String,
         val difficulty: LessonDifficulty
     ) : LessonDraft(id, name)
 
     data class StepByStep(
-        override val id: Long,
+        override val id: Int,
         override val name: String,
         val description: String,
         val difficulty: LessonDifficulty
     ) : LessonDraft(id, name)
 
     data class Test(
-        override val id: Long,
+        override val id: Int,
         override val name: String
     ) : LessonDraft(id, name)
