+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/lesson/domain/usecase/GetSectionLessonsDraftsUseCaseImpl.kt
@@ -9,7 +9,7 @@ class GetSectionLessonsDraftsUseCaseImpl(
     private val lessonRepository: LessonRepository,
     private val courseContext: CourseContext
 ) : GetSectionLessonsDraftsUseCase {
-    override suspend fun execute(sectionId: Long): List<LessonDraft> {
+    override suspend fun execute(sectionId: Int): List<LessonDraft> {
         return lessonRepository
             .getLessons(courseCode = courseContext.getCourseCode(), sectionId = sectionId)
             .sortedBy { it.orderNum }
