+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/sectionEditing/data/local/dao/SectionDraftDao.kt
@@ -9,7 +9,7 @@ import cz.cvut.docta.sectionEditing.data.model.SectionDraftEntity
 interface SectionDraftDao {
 
     @Query("SELECT * FROM section_draft WHERE id = :id")
-    suspend fun getSectionDraft(id: Long): SectionDraftEntity?
+    suspend fun getSectionDraft(id: Int): SectionDraftEntity?
 
     @Query("SELECT * FROM section_draft WHERE courseCode = :courseCode")
     suspend fun getSectionDrafts(courseCode: String): List<SectionDraftEntity>
