+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/sectionEditing/data/repository/SectionDraftRepositoryImpl.kt
@@ -7,7 +7,7 @@ class SectionDraftRepositoryImpl(
     private val localSource: SectionDraftLocalDataSource,
 ) : SectionDraftRepository {
 
-    override suspend fun getSectionDraft(id: Long): SectionDraftEntity? {
+    override suspend fun getSectionDraft(id: Int): SectionDraftEntity? {
         return localSource.getSectionDraft(id = id)
     }
